Bugfix; Added abcsmc suppor to main sim program
Bugfix was with regard to treating node ids as vector indeces.  That's often okay, as they're
assign as sequential integers, but in this case it's generally not safe, as the ebola ring
generator tends to not yield node ids that are sequential.  Error was causing seg faults or
other errors in unrelated places