Refactored & generalized ring generator; at least 1 bugfix
Generator now allows generation of rings with n levels, rather than
strictly index case (level 0), level 1 and level 2.

Previously, index case had an expected degree of K, and because of the
connection from the inside of the ring outward, other nodes ended up
with an expected degree of K + 1.

Note that the nodes in the outermost level end up with a degree less than
K because, by definition, they are not completely wired--otherwise, they
wouldn't be the outermost level.