Ring generator bugfix; some refactoring/renaming
Previous bugfix re: wiring A->B and B->A was a bit o'er zealous.
That should only apply to potential connections within a level, not
from the current outer-most level to disconnected nodes.

We're now using "ring" to mean the whole network, and "level" to
mean a set of nodes that are all the same number of steps removed
from the index case.  This is for consistency with the rest of the
world, even though it is not my prefered nomenclature.