Critical bug-fix to network generator
Changes introduced in commit bec8591c938bc900e246cbc4b03835a572490379
resulted in the wiring probabilities matrix being incompletely populated.
Specifically, only one triangle of the matrix.  The other triangle is
what was checked when considering edges within a level.  The result is that
cycles of >3 nodes were possible, but triangles were not.

This commit causes both triangles of the matrix to be populated.